feat: shrink decision mockData and route ops UI through v1 client
Extract decision/ops mocks into mockData.decisions.ts and move decision
surfaces onto api/v1/decisions so ops UI stops depending on the god barrel.
