+++ frontend/src/components/DecisionInspector.test.tsx
@@ -10,8 +10,8 @@ const { getAudit } = vi.hoisted(() => ({
   getAudit: vi.fn(),
 }));
 
-vi.mock("@/api/client", async (importOriginal) => {
-  const actual = await importOriginal<typeof import("@/api/client")>();
+vi.mock("@/api/v1/decisions", async (importOriginal) => {
+  const actual = await importOriginal<typeof import("@/api/v1/decisions")>();
   return {
     ...actual,
     decisions: {
